Only use mvnw on win32 if JAVA_HOME set
The mvnw.cmd script requires that JAVA_HOME be set, it does not try to
automatically determine it like the mvnw shell script does.  Before
returning mvnw on win32, make sure it exists _and_ JAVA_HOME is set.
Add tests for both types of platforms.

Rename file containing determineMavenCommand to mavenCommand since it
should not have a class name if it does not contain one class having
the same name.

Closes #62